replace standard serialization implementation with async
