+++ src/lib.rs
@@ -73,7 +73,6 @@ macro_rules! capnp_word {
 }
 
 pub mod any_pointer;
-pub mod async;
 pub mod capability;
 pub mod data;
 pub mod data_list;
